move docker types to docker domain
