+++ server/src/types/network.ts
@@ -9,6 +9,7 @@ import type {
 	GotTimings,
 	ILighthouseAudit,
 } from "@/domain/checks/check.type.js";
+import { DockerContainerInfo, DockerContainerLogs, DockerContainerSummary } from "@/domain/docker/docker.type.js";
 import type { DnsRecordType, Monitor, MonitorMatchMethod, MonitorStatus, MonitorType } from "@/domain/monitors/monitor.type.js";
 
 import type { QueryResult } from "gamedig";
@@ -97,68 +98,6 @@ export interface HardwareStatusPayload {
 }
 
 // Docker host monitoring
-export const DockerContainerStates = ["created", "running", "paused", "restarting", "removing", "exited", "dead"] as const;
-export type DockerContainerState = (typeof DockerContainerStates)[number];
-
-export const DockerHealthStatuses = ["healthy", "unhealthy", "starting", "none"] as const;
-export type DockerHealthStatus = (typeof DockerHealthStatuses)[number];
-
-export const DockerLogStreams = ["stdout", "stderr"] as const;
-export type DockerLogStream = (typeof DockerLogStreams)[number];
-export interface DockerLogLine {
-	ts: string;
-	stream: DockerLogStream;
-	text: string;
-}
-
-export interface DockerContainerLogs {
-	containerId: string;
-	containerName: string;
-	lines: DockerLogLine[];
-}
-
-export interface DockerContainerSummary {
-	total: number;
-	running: number;
-	stopped: number;
-	unhealthy: number;
-}
-
-export interface DockerContainerInfo {
-	id: string;
-	name: string;
-	image: string;
-	state: DockerContainerState;
-	status: string;
-	health: DockerHealthStatus;
-	cpuPct?: number; // fraction of one core (docker stats convention / 100); exceeds 1 when using multiple cores
-	memoryUsedBytes?: number;
-	memoryLimitBytes?: number;
-	memoryPct?: number; // 0-1 fraction of the memory limit
-	restartCount?: number;
-	startedAt?: string; // ISO date
-	ports?: DockerContainerPort[];
-	mounts?: DockerContainerMount[];
-}
-
-export const DockerPortProtocols = ["tcp", "udp", "sctp"] as const;
-export type DockerPortProtocol = (typeof DockerPortProtocols)[number];
-
-export interface DockerContainerPort {
-	privatePort: number;
-	protocol: DockerPortProtocol;
-	publicPort?: number;
-	hostIp?: string;
-}
-
-export interface DockerContainerMount {
-	type: string; // open set: bind, volume, tmpfs, image, npipe, cluster, …
-	name?: string;
-	source: string;
-	destination: string;
-	mode: string;
-	rw: boolean;
-}
 
 export interface DockerStatusPayload {
 	containers: DockerContainerInfo[];
